Crate zerofrom

This crate provides ZeroFrom, a trait for converting types in a zero-copy way.

See the documentation of ZeroFrom for more details.

Traits

ZeroFrom

Trait for types that can be created from a reference to a different type C with no allocations, i.e. a zero-copy (zero-alloc) version of “From”

Derive Macros

ZeroFrom

Custom derive for zerofrom::ZeroFrom,